# tests/unit/data/openapi.json
{
  "components": {
    "schemas": {
...
      "CharacterExportRequest": {
        "properties": {
          "backstory": {
            "title": "Backstory",
            "type": "string"
          },
          "character_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "personality": {
            "title": "Personality",
            "type": "string"
          },
          "physical_description": {
            "title": "Physical Description",
            "type": "string"
          },
          "portrait_prompt": {
            "default": "",
            "title": "Portrait Prompt",
            "type": "string"
          },
          "save_id": {
            "default": "",
            "title": "Save Id",
            "type": "string"
          },
          "save_title": {
            "default": "",
            "title": "Save Title",
            "type": "string"
          }
        },
        "required": [
          "name",
          "backstory",
          "personality",
          "physical_description"
        ],
        "title": "CharacterExportRequest",
        "type": "object"
      },
...
      "CharacterLibraryResponse": {
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterLibraryEntry"
            },
            "title": "Characters",
            "type": "array"
          }
        },
        "required": [
          "characters"
        ],
        "title": "CharacterLibraryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "par-storygen API"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/characters": {
      "get": {
        "description": "List all library characters.",
        "operationId": "list_characters_api_characters_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Characters",
        "tags": [
          "characters"
        ]
      },
      "post": {
        "description": "Export a character to the library.",
        "operationId": "export_character_api_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Character",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/create": {
      "post": {
        "description": "Create a new library character via LLM from a concept description.",
        "operationId": "create_character_api_characters_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Character",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/import-from-story": {
      "post": {
        "description": "Import characters from a saved game into the library.",
        "operationId": "import_from_story_api_characters_import_from_story_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharacterLibraryEntry"
                  },
                  "title": "Response Import From Story Api Characters Import From Story Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import From Story",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/{library_id}": {
      "delete": {
        "description": "Remove a character from the library.",
        "operationId": "delete_character_api_characters__library_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Character",
        "tags": [
          "characters"
        ]
      },
      "get": {
        "description": "Get a single library character.",
        "operationId": "get_character_api_characters__library_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character",
        "tags": [
          "characters"
        ]
      },
      "put": {
        "description": "Edit character fields.",
        "operationId": "update_character_api_characters__library_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Character",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/{library_id}/edit-portrait": {
      "post": {
        "description": "Edit portrait prompt and regenerate.",
        "operationId": "edit_portrait_api_characters__library_id__edit_portrait_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortraitEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Portrait",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/{library_id}/portrait": {
      "get": {
        "description": "Serve a library character's portrait PNG.",
        "operationId": "get_character_portrait_api_characters__library_id__portrait_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character Portrait",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/{library_id}/reference-image": {
      "delete": {
        "description": "Remove a character's reference image.",
        "operationId": "delete_reference_image_api_characters__library_id__reference_image_delete",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Reference Image",
        "tags": [
          "characters"
        ]
      },
      "get": {
        "description": "Serve a library character's reference image PNG.",
        "operationId": "get_reference_image_api_characters__library_id__reference_image_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reference Image",
        "tags": [
          "characters"
        ]
      },
      "post": {
        "description": "Upload a reference image for a library character.\n\nAccepts multipart form with ``image`` file and ``mode`` field\n(``\"use_as_is\"`` or ``\"style_transfer\"``).",
        "operationId": "upload_reference_image_api_characters__library_id__reference_image_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "style_transfer",
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_reference_image_api_characters__library_id__reference_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLibraryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Reference Image",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/wizard/characters": {
      "post": {
        "description": "Generate characters for a theme.",
        "operationId": "wizard_characters_api_wizard_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WizardCharactersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WizardCharactersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wizard Characters",
        "tags": [
          "wizard"
        ]
      }
    },
...
  }
}